requirements.txt
ablog==0.11.11
accessible-pygments==0.0.5
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
alabaster==0.7.16
albumentations==1.0.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astroid==3.3.8
asttokens==3.0.0
async-lru==2.0.4
attrs==25.1.0
azure-ai-documentintelligence==1.0.0
azure-core==1.32.0
azure-identity==1.20.0
babel==2.16.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
black==24.8.0
bleach==6.2.0
blinker==1.9.0
blp==0.0.3
Brotli @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_f7i0oxypt6/croot/brotli-split_1736182464088/work
cachetools==5.5.2
certifi @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_53iribkvo4/croot/certifi_1738623743875/work/certifi
cffi==1.17.1
charset-normalizer==3.4.1
chartbook @ git+https://github.com/jmbejara/chartbook@9ddc1719efaf96d24fe9eec189dc993f73f9b4e8
clerk-backend-api==1.8.0
click==8.1.8
cloudpickle==3.1.1
cobble==0.1.4
colorama==0.4.6
comm==0.2.2
config==0.5.1
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
Cython==3.0.11
dataparser==0.0.2
datasets==3.3.2
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.8
distro==1.9.0
dnspython==2.7.0
docutils==0.20.1
doit==0.36.0
email_validator==2.2.0
et_xmlfile==2.0.0
eval_type_backport==0.2.2
exchange_calendars==4.8
executing==2.2.0
fabric==3.2.2
fastapi==0.115.8
fastapi-cli==0.0.7
fastjsonschema==2.21.1
feedgen==1.0.0
filelock==3.17.0
fonttools==4.55.6
formulaic==1.1.1
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
holidays==0.65
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
imageio==2.37.0
imagesize==1.4.1
importlib_metadata==8.6.1
inflection==0.5.1
iniconfig==2.0.0
interface-meta==1.3.0
invoke==2.2.0
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
isodate==0.7.2
isoduration==20.11.0
JayDeBeApi==1.2.3
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jpype1==1.5.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-book==1.0.3
jupyter-cache==1.0.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
korean-lunar-calendar==0.3.1
latexcodec==3.0.0
lazy_loader==0.4
Levenshtein==0.26.1
lightning==2.5.0.post0
lightning-utilities==0.12.0
linearmodels==6.1
linkify-it-py==2.0.3
lxml @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_1cc0pw19wu/croot/lxml_1737040416995/work
mammoth==1.9.0
markdown-it-py==3.0.0
markdownify==1.0.0
markitdown==0.0.1a4
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistune==3.1.0
more-itertools==10.6.0
mpmath==1.3.0
msal==1.31.1
msal-extensions==1.2.0
multidict==6.1.0
multiprocess==0.70.16
multitasking==0.0.11
munch==4.0.0
mypy-extensions==1.0.0
myst-nb==1.1.2
myst-parser==2.0.0
narwhals==1.28.0
Nasdaq-Data-Link==1.0.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
nougat-ocr==0.1.17
numexpr @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_b3kvvt6tc6/croot/numexpr_1730215947700/work
numpy==1.26.4
numpydoc==1.8.0
olefile==0.47
openai==1.64.0
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.10.15
overrides==7.7.0
packaging==23.2
pandas @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_4aifrweohv/croot/pandas_1732735109535/work/dist/pandas-2.2.3-cp312-cp312-macosx_11_0_arm64.whl#sha256=123f63647068e74a9e99f9fd6ab9ee71cd9f6d1c55fa82c75a587a95021910a5
pandas-datareader @ file:///tmp/build/80754af9/pandas-datareader_1626259591421/work
pandas_market_calendars==4.4.1
pandocfilters==1.5.1
paramiko==3.5.0
parso==0.8.4
pathspec==0.12.1
pathvalidate==3.2.3
patsy==1.0.1
pdfminer.six==20240706
peewee==3.17.8
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
polars==1.9.0
portalocker==2.10.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.3
psutil==6.1.1
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.28
pyarrow==19.0.0
pybtex==0.24.0
pybtex-docutils==1.0.3
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydata-sphinx-theme==0.15.4
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
pyhdfe==0.2.0
PyJWT==2.10.1
pyluach==2.2.0
PyNaCl==1.5.0
pyparsing==3.2.1
pypdf==5.3.0
pypdfium2==4.30.1
PySocks @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/pysocks_1699237568675/work
pytest==8.3.3
python-dateutil @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_66ud1l42_h/croot/python-dateutil_1716495741162/work
python-decouple==3.8
python-dotenv==1.0.1
python-json-logger==3.2.1
python-Levenshtein==0.26.1
python-multipart==0.0.20
python-pptx==1.0.2
pytorch-lightning==2.5.0.post0
pytz @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_a4b76c83ik/croot/pytz_1713974318928/work
pyxlsb==1.0.10
PyYAML==6.0.2
pyzmq==26.2.0
Quandl==3.7.0
QuantLib==1.37
QuantStats==0.0.64
quantstats-lumi @ git+https://github.com/jmbejara/quantstats_lumi@6b765f0f464bd0bdecddbaa1ca28cf4e58c20969
RapidFuzz==3.12.1
referencing==0.36.2
regex==2024.11.6
requests @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_ee45nsd33z/croot/requests_1730999134038/work
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.22.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
ruff==0.9.3
safetensors==0.5.3
scikit-fuzzy==0.5.0
scikit-image==0.25.2
scikit-learn==1.5.2
scipy==1.12.0
sconf==0.2.5
seaborn==0.13.2
Send2Trash==1.8.3
sentencepiece==0.2.0
setuptools==75.8.0
setuptools-scm==8.1.0
shellingham==1.5.4
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
SpeechRecognition==3.14.1
Sphinx==7.3.7
sphinx-autodoc2==0.5.0
sphinx-book-theme==1.1.3
sphinx-comments==0.0.3
sphinx-copybutton==0.5.2
sphinx-jupyterbook-latex==1.0.0
sphinx-multitoc-numbering==0.1.3
sphinx-thebe==0.3.1
sphinx-togglebutton==0.3.2
sphinx_design==0.6.1
sphinx_external_toc==1.0.1
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-bibtex==2.6.3
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sphinxext-opengraph==0.9.1
sphinxext-rediraffe==0.2.7
SQLAlchemy==2.0.37
stack-data==0.6.3
starlette==0.45.3
statsmodels==0.14.4
streamlit==1.42.2
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2025.2.18
timm==0.5.4
tinycss2==1.4.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
tomli_w==1.2.0
toolz==1.0.0
torch==2.6.0
torchmetrics==1.6.1
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata @ file:///croot/python-tzdata_1690578112552/work
uc-micro-py==1.0.3
uri-template==1.3.0
urllib3 @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_8dwi5l2dj0/croot/urllib3_1737133640453/work
uvicorn==0.34.0
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0
weightedstats==0.4.1
wheel==0.44.0
widgetsnbextension==4.0.13
wrapt==1.17.2
wrds==3.2.0
xbbg==0.7.7
xlrd==2.0.1
XlsxWriter==3.2.2
xlwings==0.33.3
xxhash==3.5.0
yarl==1.18.3
yfinance==0.2.52
youtube-transcript-api==0.6.3
zipp==3.21.0
zstandard==0.23.0
